+++ src/main/kotlin/org/simplemc/simplehealthbars2/healthbar/NameHealthbar.kt
@@ -1,23 +1,19 @@
 package org.simplemc.simplehealthbars2.healthbar
 
 import org.bukkit.entity.LivingEntity
-import org.simplemc.simplehealthbars2.getCustomDisplayName
-import org.simplemc.simplehealthbars2.setCustomDisplayName
+import org.simplemc.simplehealthbars2.getDisplayName
 
 class NameHealthbar(config: Config) : StringHealthbar(config), MobHealthbar {
     override fun updateHealth(source: LivingEntity?, target: LivingEntity, damage: Double): (() -> Unit) {
-        val hadCustomName = target.isCustomNameVisible
-        val oldName = target.getCustomDisplayName()
+        val oldCustomName = target.customName
+        val wasCustomNameVisible = target.isCustomNameVisible
 
-        target.setCustomDisplayName(formatHealthbar(target, oldName, damage))
+        target.customName = formatHealthbar(target, target.getDisplayName(), damage)
+        target.isCustomNameVisible = true
 
         return {
-            if (hadCustomName) {
-                target.setCustomDisplayName(oldName)
-            } else {
-                target.isCustomNameVisible = false
-                target.customName = null
-            }
+            target.customName = oldCustomName
+            target.isCustomNameVisible = wasCustomNameVisible
         }
     }
 }