+++ src/main/kotlin/org/simplemc/simplehealthbars2/healthbar/StringHealthbar.kt
@@ -13,7 +13,7 @@ abstract class StringHealthbar(final override val config: Config) : Healthbar {
         override val duration: Duration = Duration.ofSeconds(5),
         val length: Int = 20,
         val char: Char = 0x25ae.toChar(),
-        val showMobNames: Boolean = true
+        val showMobNames: Boolean = true,
     ) : Healthbar.Config
 
     private val step = config.char.toString()